\name{stastools-package}
\alias{stastools-package}
\alias{stastools}
\docType{package}
\title{
\packageTitle{stastools}
}
\description{
\packageDescription{stastools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stastools}
\packageIndices{stastools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{stastools}

Maintainer: \packageMaintainer{stastools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
